backend/requirement.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.0.post1
